requirements.txt
# Previewless Insight Viewer - Production Dependencies
# Python 3.10+

# UI Framework
PySide6>=6.6.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# Database
# SQLite is built into Python, but we may want better tooling
# (optional) pysqlite3>=0.5.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.2

# Ollama client (lightweight HTTP client)
requests>=2.31.0
